Add pause menu and a starting music manager
We create a pause menu when pressing esc while in game with return, config and back to main menu.

Music had to be handled specially rather than just paused to allow for changing music on the same resource.

A basic api has been created to stop and change the song- it may change later to allow for fadeins and fadeouts.